# Add benchmarks for the lazily parsed data