Also disable parallel codegen for benchmarks. This improves the benchmarks due to better inlining, and makes them a more reliable tool for determining the effect of code changes.

Before:
```
test context_new_and_query_location       ... bench:   3,111,057 ns/iter (+/- 530,666)
test context_new_and_query_with_functions ... bench:   4,172,368 ns/iter (+/- 556,175)
test context_new_location                 ... bench:   2,857,732 ns/iter (+/- 152,999)
test context_query_location               ... bench:     713,752 ns/iter (+/- 45,922)
test context_query_with_functions         ... bench:   4,526,699 ns/iter (+/- 567,636)
```

After:
```
test context_new_and_query_location       ... bench:   2,953,517 ns/iter (+/- 103,705)
test context_new_and_query_with_functions ... bench:   3,803,000 ns/iter (+/- 610,173)
test context_new_location                 ... bench:   2,784,894 ns/iter (+/- 694,324)
test context_query_location               ... bench:     717,903 ns/iter (+/- 62,360)
test context_query_with_functions         ... bench:   4,258,835 ns/iter (+/- 171,619)
```